+++ src/chains/definitions/lightlinkPhoenix.ts
@@ -20,5 +20,11 @@ export const lightlinkPhoenix = /*#__PURE__*/ defineChain({
       url: 'https://phoenix.lightlink.io',
     },
   },
+  contracts: {
+    multicall3: {
+      address: '0xca11bde05977b3631167028862be2a173976ca11',
+      blockCreated: 125_499_184,
+    },
+  },
   testnet: false,
 })