+++ src/chains/definitions/xrOne.ts
@@ -5,16 +5,19 @@ export const xrOne = /*#__PURE__*/ defineChain({
   name: 'XR One',
   nativeCurrency: {
     decimals: 18,
-    name: 'XR',
-    symbol: 'XR',
+    name: 'XR1',
+    symbol: 'XR1',
   },
   rpcUrls: {
-    default: { http: ['https://xr-one.calderachain.xyz/http'] },
+    default: {
+      http: ['https://xr1.calderachain.xyz/http'],
+      webSocket: ['wss://xr1.calderachain.xyz/ws'],
+    },
   },
   blockExplorers: {
     default: {
       name: 'Blockscout',
-      url: 'https://xr-one.calderaexplorer.xyz',
+      url: 'https://xr1.calderaexplorer.xyz',
     },
   },
   testnet: false,