+++ src/experimental/erc7821/index.ts
@@ -1,4 +1,3 @@
-/**  */
 // biome-ignore lint/performance/noBarrelFile: entrypoint
 export {
   type ExecuteErrorType,
@@ -6,6 +5,12 @@ export {
   type ExecuteReturnType,
   execute,
 } from './actions/execute.js'
+export {
+  type ExecuteBatchesErrorType,
+  type ExecuteBatchesParameters,
+  type ExecuteBatchesReturnType,
+  executeBatches,
+} from './actions/executeBatches.js'
 export {
   type SupportsExecutionModeErrorType,
   type SupportsExecutionModeParameters,
@@ -21,3 +26,25 @@ export {
 } from './errors.js'
 
 export { type Erc7821Actions, erc7821Actions } from './decorators/erc7821.js'
+
+export {
+  type EncodeCallsErrorType,
+  encodeCalls,
+} from './utils/encodeCalls.js'
+export {
+  type EncodeExecuteBatchesDataErrorType,
+  type EncodeExecuteBatchesDataParameters,
+  type EncodeExecuteBatchesDataReturnType,
+  encodeExecuteBatchesData,
+} from './utils/encodeExecuteBatchesData.js'
+export {
+  type EncodeExecuteDataErrorType,
+  type EncodeExecuteDataParameters,
+  type EncodeExecuteDataReturnType,
+  encodeExecuteData,
+} from './utils/encodeExecuteData.js'
+export {
+  type GetExecuteErrorParameters,
+  type GetExecuteErrorReturnType,
+  getExecuteError,
+} from './utils/getExecuteError.js'